+++ project/src/main/java/lu/uni/adtool/domains/NodeRanker.java
@@ -1,0 +1,8 @@
+package lu.uni.adtool.domains;
+
+import lu.uni.adtool.domains.rings.Ring;
+import lu.uni.adtool.tree.Node;
+
+public interface NodeRanker {
+  void rankNode(Node node, Ring value);
+}