+++ project/src/main/java/lu/uni/adtool/tree/SandNode.java
@@ -1,9 +1,8 @@
 package lu.uni.adtool.tree;
 
-import lu.uni.adtool.domains.SandDomain;
+import lu.uni.adtool.domains.RankExporter;
 import lu.uni.adtool.domains.ValuationDomain;
 import lu.uni.adtool.domains.rings.Ring;
-import lu.uni.adtool.tools.Debug;
 import lu.uni.adtool.tools.Options;
 
 import java.io.DataInputStream;
@@ -79,7 +78,7 @@ public class SandNode extends GuiNode {
    * Export to XML using format used by the first version of ADTool
    *
    */
-  public XElement exportXml(ArrayList<ValuationDomain> domains) {
+  public XElement exportXml(ArrayList<ValuationDomain> domains, ArrayList<RankExporter> rankers) {
     XElement result = new XElement("node");
     result.addString("refinement", typeToXml(type));
     result.addElement("label").setString(getName());
@@ -103,11 +102,21 @@ public class SandNode extends GuiNode {
             param.setString(vd.getTermValue(this).toString());
           }
         }
+        if (rankers != null && rankers.size() > i) {
+          for (int j=0; j < Options.rank_noRanked; j++) {
+            Ring value = rankers.get(i).getValue(this, j);
+            if(value != null) {
+              XElement rank = result.addElement("ranking");
+              rank.addInt("rank", j + 1);
+              rank.setString(value.toString());
+            }
+          }
+        }
       }
     }
     if (this.getChildren() != null) {
       for (Node node : this.getNotNullChildren()) {
-        result.addElement(((SandNode) node).exportXml(domains));
+        result.addElement(((SandNode) node).exportXml(domains, rankers));
       }
     }
     return result;