+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the ADTool will be documented in this file.
 - fixed not updating node size after checking "Computed Values" checkbox
 - added copy/paste of values in Valuations View
 - added command line options
+- fixed several bugs in wrong/missing messages
+- fixed bug when marking ranking on countered node with no children
 
 ## [2.0.1] - 2016-03-01
 - fixed bug with wrong Sand Tree terms